# GolemClientMockAPI/Entities/Proposal.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace GolemClientMockAPI.Entities
{
    public class Proposal
    {
        public int InternalId { get; set; }

    }
}